keras/requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
funcsigs==1.0.2
h5py==2.7.0
html5lib==0.9999999
Keras==1.1.2
Markdown==2.6.8
mock==2.0.0
olefile==0.44
pandas==0.20.3
pbr==3.1.1
Pillow==4.2.1
protobuf==3.3.0
PyYAML==3.12
scikit-learn==0.19.0
scipy==0.19.1
six==1.10.0
tensorflow==0.12.1
Theano==0.9.0
Werkzeug==0.12.2
